requirements.txt
torch 
torchvision 
torchaudio
torchsummary
numpy
fonttools
tqdm
matplotlib
imageio
Pillow

# pip install -r requirements.txt